[chromedriver] Add RemoveVirtualAuthenticator command
Add the `RemoveVirtualAuthenticator` command to ChromeDriver.

A draft of the proposed addition to the specification is available at
https://github.com/w3c/webauthn/pull/1256

This is one in a series of patches intended to create a Testing API for
WebAuthn, for use in Web Platform Tests and by external webauthn tests.

For an overview of overall design, please see
https://docs.google.com/document/d/1bp2cMgjm2HSpvL9-WsJoIQMsBi1oKGQY6CvWD-9WmIQ/edit?usp=sharing

Bug: 922572
Change-Id: I5b6094cc5eb9a52afc8ee89d548c3452a728b1fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1713301
Auto-Submit: Nina Satragno <nsatragno@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679727}