Get rid of masking logic from WebSocketFrameParser
We are planning to remove memcpy in
WebSocketFrameParser::DecodeFramePayload. Unfortunately, we cannot
keep the constness of |payload| as long as we have the unmasking logic
there.

This CL moves the logic to the caller. It is a spec violation for a
server to set the "masked" bit in a frame, so we don't need the
additional logic in net/websockets.

Bug: 865001
Change-Id: I73042683fba799d5b4072ec619ab52ece6723e17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1784244
Commit-Queue: Yoichi Osato <yoichio@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693440}