Set async dispatch event flag in ExecutePerformActions
ExecutePerformActions function is called from testdriver Action API and
Webdriver API, and most of cases users assume that we are sending the
events synchronously. In order to make all user cases work, we added a
flag in action to indicate if we want to dispatch events synchronously
or asynchronously. The default way is synchronously.

Bug: 999814, 606367
Change-Id: I41f0c17cd7b0c16e9b6c6ed54a18dce61a3be40e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1785003
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694439}