[ChromeDriver] [ChromeDriver] Wait for Current Frame to load
Reland: When Waiting for Pending Navigation, code will respond to loading events
for the current frame as well as the top frame. ChromeDriver will
therefore wait for changes in the current frame to complete before
running the next command.

Fix the check before adding a listener to the NavigationTracker. Resolves
bug with invalid static-cast.

Bug: chromedriver:3164
Change-Id: Iff935ba539e72cbd858c837ec1d4388afb1b73ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924789
Commit-Queue: Tricia Crichton <triciac@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716789}