Reformat all gn files in /chrome/test/chromedriver.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=johnchen@chromium.org

Bug: 1041419
Change-Id: Id338c2ce6f613d444df42aaaa48f538cb91d8cf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998626
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732090}