[chromedriver] Do not pass in minidump path if empty
When specify log-path, minidump path will be empty if specifying
file in current dir such as --log-path=file.log. In this case,
we don't have to pass in the minidump path option to driver as
that is mainly used on bot and it will be specified on bot.

Bug: 3437
Change-Id: Iacc53af30330bc2615499aa2667adb61fd460a43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163422
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Shengfa Lin <shengfa@google.com>
Cr-Commit-Position: refs/heads/master@{#762143}