Send related WebDriver input messages in the same websocket frame
It is currently possible for clicks sent from WebDriver to be
interleaved with real mouse events from the OS.

This change updates WebViewImpl::DispatchMouseEvents,
DispatchTouchEvents, and DispatchKeyEvents to use
SendCommandAndIgnoreResponse for all but the last message.

This group all of the requested input events into a single websocket
frame so they will be processed together on the browser's UI thread.
The last event in an input sequence still uses SendCommand so that the
WebDriver command will not return until the input sequence is complete.
The methods above and various helper methods have been updated to use
std::vector instead of std::list so that the code can quickly check
which event is the last one in a sequence.

Bug: 1081362
Change-Id: I67f4e2b0fd8c7986e67bbc305a9bb0669bca3467
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194904
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Lan Wei <lanwei@chromium.org>
Commit-Queue: Brandon Walderman <brwalder@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#767913}