[chromedriver] Properly set --enable-logging when Chrome logs are enabled
Previously the code would default to always passing --enable-logging (no
parameter) to Chrome. This only logs to stderr on Debug Chrome builds or
when running in headless mode[0], so in other configurations the user
would also be required to set '--enable-logging=stderr' as an args
capability. Instead, use the new '--enable-chrome-logs' flag as a signal
that the user wants browser logs sent to stderr, and set either
'--enable-logging' or '--enable-logging=stderr' accordingly.

[0]: https://docs.google.com/document/d/1DsvGXPri3exlQupj3d5IbOvI1tzIw2ePWbePDZVp1Cc/edit

Bug: 1109233
Change-Id: Ife69170e67746b5ba6cd8dbd0ddb8041827d76c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2341891
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799688}