[chromedriver] Fix passing deferencing nullptr value
DevTools event listener is passing a dereferenced pointer value
to navigation. However, sometimes it's dereferencing a nullptr value
(from an uninitialized unique_ptr)
causing navigation tracker to crash when try to use the reference.
See bug list 3578 for a case when that pointer is nullptr.
When navigating to a blank page and an alert is triggered,
ChromeDriver will not wait for the original navigation to
finish and set the original Page.navigate response info state
to ignore; when the response finally come back, ChromeDriver is
not copying the result in the response causing the result uninitialized.

Bug: chromedriver:3578,chromedriver:3488
Change-Id: I7fd6159d724036a7354a47844da9d14d76c70b3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2370091
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Shengfa Lin <shengfa@google.com>
Cr-Commit-Position: refs/heads/master@{#801149}