Roll catapult to 532e564 and update for changes in it.
This CL rolls //third_party/catapult to
532e564f622742fc93a7f91dd3556d711e778a7e and picks up

https://chromium-review.googlesource.com/c/catapult/+/2450450

which removes zipfile_2_7_13.py as a dependency. This CL
also updates the corresponding Chromium .pydeps files that mentioned
it.

That change also added a dependency on `six`, which seems to be
pervasively assumed in other parts of the catapult codebase as
present in the environment; however, it isn't guaranteed to be
there in the chromium compile-time python build, so this adds
that as needed as well.

TBR=bsheedy@chromium.org

Bug: 1112471
Change-Id: I849355a5bdc18847dcbf39f3d3a9e9fbf428e3be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2449938
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#814062}